backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
supabase==2.9.1
anthropic==0.34.2
ofxparse==0.21
python-multipart==0.0.12
upstash-redis==1.3.0
redis==5.0.8
python-dotenv==1.0.1
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.5.2
